+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/UnitTests.hs
@@ -578,7 +578,6 @@ aliceStakePoolParamsSmallCost =
             { pmUrl = fromJust $ textToUrl 64 "alice.pool"
             , pmHash = byteArrayFromShortByteString "{}"
             }
-    , sppBlsKey = SNothing
     }
   where
     vkVrf = vrfVerKey $ mkVRFKeyPair @MockCrypto (RawSeed 0 0 0 0 2)