+++ libs/cardano-ledger-core/test/Test/Cardano/Ledger/State/StakePoolSpec.hs
@@ -6,7 +6,6 @@
 
 module Test.Cardano.Ledger.State.StakePoolSpec (spec) where
 
-import Cardano.Ledger.BaseTypes (EpochNo)
 import Cardano.Ledger.Coin
 import Cardano.Ledger.Core
 import Cardano.Ledger.Credential (Credential)
@@ -20,12 +19,11 @@ spec = do
   describe "StakePoolState" $ do
     prop "mkStakePoolState/stakePoolStateToPoolParams round-trip" $
       \( stakePoolParams :: StakePoolParams
-         , epochNo :: EpochNo
          , deposit :: CompactForm Coin
          , delegs :: Set (Credential Staking)
          ) ->
           let poolId = sppId stakePoolParams
               network = aaNetworkId $ sppAccountAddress stakePoolParams
-              stakePoolState = mkStakePoolState epochNo deposit delegs stakePoolParams
+              stakePoolState = mkStakePoolState deposit delegs stakePoolParams
               stakePoolParams' = stakePoolStateToStakePoolParams network poolId stakePoolState
            in stakePoolParams === stakePoolParams'